feat: search caching; log to file; fixes
Search caching:

To optimize performance of the "search" type, a cache has been added
Performing string comparison for each search request is quite heavy.
A caching mechanism has been implemented to mitigate that "issue".

Log to file:

Instead of writing messages to stdout, a log file can be specified for logging

Fixes:

The rate limit cleanup did only run once, not in a loop. This is fixed now.

Signed-off-by: moson-mo <mo-son@mailbox.org>