feat: improve performance of suggest type
We organize data for suggest lookups into "buckets".
Thus we have to process way less data when doing the matching.

With this, we greatly improve performance.
Depending on the first character of the suggest-term,
it can be up to 10 times faster than before.

We also switch back to the max limit of 20 results per query (from 100).

Signed-off-by: moson-mo <mo-son@mailbox.org>