fix: check files modified date when loading from disk
When we load a file from disk we should check the modified date
and skip reloading data when the file has not been modified.

With this we can set short intervals for reloading data,
staying closer to the time when data was exported.

The "rpc_data_last_refresh" metric is also updated with that value.
Instead of showing the date when the data was reloaded,
we show the date/time when the metadata file was generated.
This is valid for loading from a URL too.

Signed-off-by: moson-mo <mo-son@mailbox.org>